Move cluster interface to its own file
