+++ acceptance/framework/consul/cluster.go
@@ -1,0 +1,18 @@
+package consul
+
+import (
+	"testing"
+
+	"github.com/hashicorp/consul/api"
+)
+
+// Cluster represents a consul cluster object
+type Cluster interface {
+	Create(t *testing.T)
+	Destroy(t *testing.T)
+	// Upgrade runs helm upgrade. It will merge the helm values from the
+	// initial install with helmValues. Any keys that were previously set
+	// will be overridden by the helmValues keys.
+	Upgrade(t *testing.T, helmValues map[string]string)
+	SetupConsulClient(t *testing.T, secure bool) *api.Client
+}
